+++ temba/orgs/models.py
@@ -279,14 +279,11 @@ class User(AuthUser):
 
         return role.has_perm(permission)
 
-    def get_api_token(self, org) -> str:
-        from temba.api.models import APIToken
-
-        try:
-            token = APIToken.get_or_create(org, self)
-            return token.key
-        except ValueError:
-            return None
+    def get_api_tokens(self, org):
+        """
+        Gets this users active API tokens for the given org
+        """
+        return self.api_tokens.filter(org=org, is_active=True)
 
     def as_engine_ref(self) -> dict:
         return {"email": self.email, "name": self.name}
@@ -303,6 +300,9 @@ class User(AuthUser):
         self.password = ""
         self.is_active = False
         self.save()
+
+        # release any API tokens
+        self.api_tokens.update(is_active=False)
 
         # release any orgs we own
         for org in self.get_owned_orgs():